Add _valgrind.callgrind_dump_stats_at to dump callgrind profile data from managed code
Django goes through multiple stages, each of which should be inspected separately for performance measurement.

This change adds `_valgrind.callgrind_dump_stats_at` to dump the collected callgrind statistics at various program points and reset them.

A concrete usage of this function is to call it once after a Django process gets initialized and get one or two requests to dump all perf stats for the startup measurement, and uses it once again after sending a meaningful number of requests to measure the workload of handling actual requests.

This change also adds header files for other valgrind tools that work in a similar fashion and `_valgrind` module so that we can add more valgrind hooks of the same kind in the future easily.

Based on Facebook D17882186