Declare all built-in types in python
Add code that checks that all built-in types have been declared with a `class xxx(bootstrap=True)` declaration in a python file. This ensures that the types are properly initialized. Without this we would miss details like automatically generated `__dict__` or `__doc__` on those types.

Added missing declarations to `builtins.py`.

Based on Facebook D17908838