Move _builtins._traceback to builtins._traceback
I could not find a compelling reason why the `_traceback` type needs to be in the `_builtins` module rather than `builtins` that contains all other builtin types.

This leaves `_builtins` without any types. That helps an upcoming commit that forces every type to be declared via `class xxx(bootrstrap=True)` in a python file. We cannot use such declarations in `_builtins` because `__build_class__` does not work properly that early in the bootstrap yet.

Based on Facebook D17908745