Import modules used in builtins.py outside functions
Executing "IMPORT_NAME" opcode for `import XXX` is a very expensive operation that involves a function call, a dict lookup, etc. Despite the cost, importing on-demand happens quite often in some crucial builtin functions.

This change performs import core modules in a new `_init` function in builtins.py, and Runtime calls it after all other core modules are processed by it.

**Perf Impact**

The perf impact of this diff was measured in terms of callgrind numbers to start Django and process the fifth request to the Django minimal example:

| Version | Startup | 5th request  |
| -- | -- | -- |
| HEAD | 69642094222 | 17175886 |
| This Diff | 69490682472 | 14053059 |

The measurement shows that this diff shows no improvement for the Django startup, and shows 19% improvement in processing the 5th request.

Based on Facebook D17916373