Explicitly mark all assembly jumps as near/far
The `near` argument of `jcc`/`jmp` defaulting to `kFarJump` makes it too easy to forget using a near jump where it would be possible. Removes the default argument and mark some jumps defaulting to far as near jumps.

Also removed the unused `jump` function (that just forwarded to `jmp`).

Based on Facebook D17954273