Compute type flags from base classes
Summary: We used to have a `// Copy down class flags from bases` comment and then the following code would go through the whole MRO instead of just the list of bases. I don't see any reason why this would be necessary, change the code to just go through the list of bases. The bases already have accumulated all the flags of their superclasses, so searching further in the MRO should not help.

Based on Facebook D17982191