Refactor code exiting from recursive interpreters
Background:
Previously to exit a recursive interpreter the `RETURN_VALUE` opcode would check whether we want to exit a recursive interpreter and if not pop the frame. This meant we did not pop the frame when exiting a recursive interpreter and the `popFrame()` call was performed by the `do_return` lambda at the end of `Interpreter::execute`.

Changes:
This change cleans up the code by unifying all `popFrame()` operations inside `handleReturn` and `unwind` and no longer having the 3rd pop operation in `do_return`.

- Changes `handleReturn`/`unwind` to call `popFrame()` when exiting a recursive interpreter.
- Change the `handleReturn` return value to communicate both a possible return value (in case we exit a recursive interpreter) or `Error::error()` to continue the interpreter loop.
- A `popFrame()` call had to be added to the yield opcodes because we are no longer calling it at the end of the interpreter function.
- Since `unwind` no longer needs to push `Error::exception()` onto the stack, we instead return it from the callers of `unwind`.

Based on Facebook D17945200