Build Type API that encapsulates attr cache invalidation logic
This diff moves all `Runtime::typeDict*` functions to type-builtins.cpp and change them to accept `type` object instead of `dict` object as its parameter. Alongside with this move, this diff enhances type attr modification APIs to trigger attr cache invalidation, which was a loophole Pyro had to modify type entries without correctly invalidating caches.

- `typeAtPut` uses a key of any type, which is not the target of our attr caching. Therefore, we add `DCHECK` to make sure that an attribute updated by `typeAtPut` is not cached anywhere.
- `typeAtPutByStr` and `typeAtPutById` both trigger cache invalidation with this change.
- `typeRemoveByStr` is added (moved from Runtime::classDelAttr) and also triggers cache invalidation.

A downside of this change is that we should create another handle for type dict inside these APIs. However, this impact is expected to be negligible and if it is noticeable we already have a plan to tackle that at T55801516. Also more optimizations will be enabled afterwards (types with only interned-str as keys).

Note that this diff is mainly for cache invalidation correctness.

Based on Facebook D17942823