Optimize Runtime::dictLookup by hash comparison before key comparison
The current implementation of `dict` wastes cycles at line 3244 by unnecessarily checking if the hash value of the current bucket is not None: if a bucket is not empty and also not a tombstone, it is always guaranteed that the hash value of that bucket is not None.

This change utilizes the hash value of the current bucket by comparing it to the need's hash value. This is more beneficial than not doing that since checking an object equality is complex and sometimes very costly.

**Perf Impact (on 5th request on django_minimal) **

| Version | Callgrind Number |
| -- | -- |
| Base | 8366184 |
| This Diff (including all diffs on stack) | 7916241 |

Improvement: +5.37%

Based on Facebook D18127482