Rename dict.__setitem__ to _dict_setitem to optimize its name lookup
This change converts `dict.__setitem__` to `_dict_setitem` to avoid type lookups to just call it as it renames exsiting `_dict_setitem` to `_capi_dict_setitem`. Knowing that dict is one of the most core data structure used very often, optimizing its symbol lookup is enough to boost the performance of pyro by a few percents.

`dict.update` will be moved in the same fashion in a follow-up diff to make the perf impact as great as expected (~10%)

**perf impact (CG per 5th request for the entire stack) **
- base: 7425949
- diff: 6972575
- improvement: +6.10%

Based on Facebook D18141938