Rename dict.get to _dict_get to optimize its name lookup
This change converts `dict.get` to `_dict_get` to avoid type lookups to just call it. Knowing that dict is one of the most core data structure used very often, optimizing its symbol lookup is enough to boost the performance of pyro by a few percents.

Will be followed by other diffs to remove unnecessary checks for `dict` and perf analysis will be done on the final diff on that stack.

Based on Facebook D18099911