Call more native object dealloc slots at Runtime shutdown
This is a trimmed-down version of D16973555, rebased on all the recent
changes and targeted at the test I added, which was the original leak that sent
me down this path a while ago. It restores some handle cleanup code that was
accidentally deleted in D17436677 and clears the module dict before running the
final GC. Our shutdown process is still far from perfect, but this is enough to
fix the new test.

This exposed an existing issue with how we store the `kBase` and `kBases`
slots. Each used to be stored as a `PyObject*` encoded in an `Int`, but because
they were initialized with borrowed references, there was nothing preventing us
from freeing the `ApiHandle*` pointed to by the slot. Rather than holding a new
reference in the slots tuple (which would add a bunch of gross refcounting
code) or special-casing this in the GC, let's store the `RawType` or `RawTuple`
directly in the slots tuple. It adds an extra few branches to
`PyType_GetSlot()` but simplifies all the other code that has to deal with
these slots.

Based on Facebook D17954880