Sort Runtime::dictLookup checks by their expected hit ratio
Runtime::dictLookup inspects buckets to see if a bucket falls into the 3 categories and the distribution of the bucket categories from running the 5th request on django_minimal is as follows:

| Category | Hit Count |
| -- | -- |
| Hash Match | 5511 |
| Empty | 3211 |
| Tombstone | 2 |

This diff reorders the checks in dictLookup by the frequency of the hit checks as observed in our profile. Note that this frequency is as we expect from a regular dict lookup assuming that our hash function spreads keys well.

**Perf Impact (CG for 5th request)**
- base: 5815148
- this diff: 5764483
- improvement : -0.87%

Based on Facebook D18176760