Allow only str keys in type dictionaries
This change enforces that keys in type dictionaries are of `str` type or its subclasses that don't override __eq__ and __hash__ as it removes `typeAt()` & `typeAtPut()` type API. With this change, Pyro's behavior diverges from CPython's and Pyro raises `TypeError` when a non-str keys are contained in `dict` to `type.__new__`. This difference is mentioned in docs/compatibility.md.

Note that this change doesn't enforce that type keys are interned strs. Such enforcement will be done in a follow-up diff.

Based on Facebook D17956896