Add bounds check to range.__len__ and define range.__bool__ and range.__reversed
In CPython, `range.__len__` should throw an `OverflowError` if the length is larger than a `ssize_t`. However, this means that certain functions (like `bool` and `reversed`) would also raise an `OverflowError`, so CPython also defines `range.__bool__` and `range.__reversed__` to ensure that the don't.

This is motivated by getting test_range to fully work

Based on Facebook D17938788