Remove includes of debugging.h and stray debugging code
Ideally, we would have a lint rule that catches includes of this header.

So as not to encourage the include of debugging.h in future diffs, I
referenced intializedDebugging with an extern pending a work around that
keeps the XCode linker from "dead stripping" symbols in debugging.cpp.

Based on Facebook D18145001