Consistently zero-initialize memory in PyType_GenericAlloc
`PyType_GenericAlloc()` was zero-initializing memory in some cases by using `calloc` and using uninitialized memory in others. This changes the code to consistently zero-initialize in all cases.

We should not need to zero initialize things at all, but currently need the behavior because otherwise we do not initialize dictionary slots of objects because of missing `__dictoffset__` support. Filed a task and added a `TODO`.

Based on Facebook D18193133