Store only interned str keys in type dictionaries
This change makes sure that only interned `str`s get stored as keys in types. Code changes are made to replace calls to `typeAtPut` for ones to `typeAtPutById` when possible. `typeAtPut` calls remain untouched in case that's impossible. Note that searching by interned str will be done in a following diff.

With this change,  `typeAtPut` is called at just 5 callsites as follows:

- typeobject.cpp:1247:  the type key is created by `Runtime::internStrFromCStr()`.
- typeobject.cpp:1267:  same as above
- typeobject.cpp:1287:  same as above
- builtins-module.cpp:238: copying dict items into type dict, and this process requires us to intern the keys from the dict.
- type-builtins.cpp:771:  Calling `type.__new__ (... dict...) ` requires a new type to be created based on the given `dict`. Here we're already making sure to raise an exception if keys in `dict` cannot be safely interned in the previous diff.

Based on Facebook D17959956