Stop scanning/editing source files with generate_cpython_sources.py
D18233418 removed the last instance where `generate_cpython_sources.py` removes a cpython function from a bigger file because of an implementation in PyRo.

We can simplify the generator now by just editing headers and only scanning the `capi` directory instead of everything in `ext`.

This change avoids the generator getting confused by some of my upcoming changes that bring in the cpython compiler code to the `ext` directory.

Based on Facebook D18231049