Pass View<byte> to marshal module
Summary: Change `Marshal::Reader` API to accept a `View<byte>` instead of a `const char*`. It was previously calling `std::strlen` on the data, but that does not work too well on binary data which can contain zero bytes. The problem wasn't noticed so far because currently nobody actually uses the `length_` field, but I plan to use it in an upcoming change.

Based on Facebook D18207997