Stop accepting empty bases lists in typeInit/typeNew
When creating a class in python you can specify an empty tuple of bases. This is interpreted as `(object,)` as a convenience to save typing/noise in the most common case. However there is little use from pushing this behavior into the lower-level type APIs, this changes the code to replace an empty tuple immediately with `Runtime::implicitBases()` at the managed/native boundary and remove special cases from the lower level type APIs.

This also changes code to set the bases field in the type early and stops passing around the bases tuple to APIs that already accept the type anyway.

Based on Facebook D18172049