Move dict functions from runtime to dict-builtins
The `Runtime::dictXXX` APIs in most cases do not access any runtime data (they only do so when allocating new memory when growing the underlying dictionary). This makes it wasteful to have them as non-static `Runtime` members for which we need to produce a runtime pointer before we can call them. This change moves them out of the runtime:

- Move the `Runtime::dictXXX` APIs out of the runtime into the `py` namespace in the `dict-builtins.h` header.
- Copy `dictLookup` as a private function to `dict-builtins.cpp` but leave the old implementation in `Runtime` as it is still used by the C-API handle table; we can clean this up in follow-up changes.

This gives a slight speedup because we do not need to produce and pass fewer `Runtime` pointers and because of the slightly specialized `dictLookup` function.

Based on Facebook D18157729