Fix refcounting bugs in extension tests
Fix a whole bunch of leaks or underflows across the C-API tests.

This diff removes `TypeExtensionApiTestNoFixture.DeallocSlotCalledDuringFinalize` and an ad-hoc handle cleanup in runtime destruction. The test will be added back in the handle GC diff (D18107743) and the ad-hoc cleanup will no longer be necessary.

Based on Facebook D18127312