Don't leak types in `ObjectExtensionApiTest`
This is similar to D18256332, but without the need for a custom
dealloc function (since they're not GC types).

`PyObject_Free()` is similar to `free()`: all it does is free the raw memory given
to it. Destroying a Python object involves much more than that, including
running arbitrary code based on the object's type. The vast majority of client
code (including tests) should not directly call `PyObject_Free()` on a
`PyObject*`; it should instead use `Py_DECREF()` to destroy any owned
references and let the decref operation destroy the object when appropriate.

The symptom of using `PyObject_Free()` here (and in D18256332) was that we were
freeing the object's memory without running its dealloc slot, causing us to
leak a reference to its type.

Based on Facebook D18272549