Return module proxy in PyModule_GetDict
Implement `PyModule_GetDict` by returning the module proxy.

This is useful for people passing the result to `PyEval_EvalCode` and similar functions. It will fail for people using `PyDict_xxx` functions on the result, but it should not be worse than before where we always hit an `UNIMPLEMENTED`.

Based on Facebook D18251903