Make libpython.a depend on all files in third-party/cpython
This works around a problem of libpython.a not getting rebuild because cmake has no insight into the `configure/make` make process invoked. This should fix most problems of cpython-tests not getting rebuild correctly.

It's not perfect, because the list of files is generated at cmake time and will not pick up files that got added since the last cmake run. But as long as we keep using the anti-pattern of invoking a different build system as a sub-step it should be good enough...

Based on Facebook D18303237