Implement signal-safe traceback printer
The existing traceback printer dumps the frames from least to most recent, but this requires operations, such as allocation, which are not async-signal-safe.

For signal handlers, we need a traceback printer that performs only async-signal-safe functions.

Implement this new traceback printer, and use it for `CHECK()` and `faulthandler.dump_traceback()`.

Based on Facebook D18272161