Fix kwarg trampoline dispatching
We had a bug where `prepareKeywordCall` would try to match `Varkeyargs` with any name in `varnames`, which would then match the name of the argumentss declared as a var arg or var kwarg.

I.e.
```
def a(*args, **kwargs):
  pass

a(args=1)
```
Would match the `args` argument as part of `func's `args` parameter

Based on Facebook D18372576