Enable one clang-tidy check in ext/Modules/_compile
I had all `clang-tidy` checks disabled in `ext/Modules/_compile` as we rather want to change the cpython code as little as possible. However it seems `clang-tidy` aborts if no checks are enabled at all and no files from other directories with checks enabled are processed at the same time.

Re-enable one check (`google-readability-todo`) so there is something enabled and remove a piece of unused code with a `TODO`.

Based on Facebook D18374516