Add _getframe_function, function.__module_object__; remove _getframe_globals, _getframe_code
Our system nowadays has the invariant that every frame has a function and that you can get the module or code via this function. Our APIs become more flexible when we make the low-level bits match the implementation. This change:

- Adds `_getframe_function()`.
- Expose `Function::moduleObject()` as a `__module_object__` attribute so it can be accessed in python.
- Replace `_getframe_code()` uses with `_getframe_function().__code__`.
- Replace `_getframe_globals()` uses with `_getframe_function().__module_object__.__dict__`.
- Replace `BuiltinsModule::globals()` with a managed implementation returning `_getframe_function(1).__module_object__.__dict__`.

This will also help me in an upcoming commit to query the code flags of a functions caller.

Based on Facebook D18373507