Move _getframe functions from sys to _builtins
They are a useful primitive to have in `builtins.py` so we should make them available in `_builtins` instead of just `sys`.

While on it remove the default parameter value: They are a internal functions that don't need to be optimized for convenience and calling with matching parameter numbers is slightly faster than filling in default parameters.

Based on Facebook D18364923