Make pyro-perf benchmark server more resilient to workload failures
Raising an exception when a benchmark fails (from D17910896) exposed
some pathological behavior in the server: the async task that processes jobs
from the queue would die while the rest of the server stayed up. The server
would keep accepting jobs but without anything running to process them, and it
was completely wedged from the client's point of view.

- Restructure the main server event setup to abort the whole server if
 `_execute_queued_runs()` or `_clean_stored_results()` ever terminates. This
  avoids the bad behavior mentioned above.
- Catch exceptions from `run_suite()` in `_execute_queued_runs()` and report
  them to the client, so a failing benchmark doesn't take down the whole
   server.
- Catch `subprocess.CalledProcessError`s in
  `PyroBenchmarkSuite.start_benchmarks()` and turn them into `RuntimeError`s
  with the failed process's stdout and stderr.

And a few bits of related cleanup while I was poking around:
- Remove a few checks of `completed_process.returncode != 0` that will
  never succeed.
- In the client, don't try to print a comparison table with only 1
  interpreter.
- Delete `run()` from `run.py` and import the nearly-identical one from
  `_tools.py`.

Based on Facebook D18400970