Monomorphic caching for LOAD_ATTR with type and module receivers
This change introduces specialized opcodes for monomorphic forms
of `LOAD_ATTR` where the expected receiver is a `type` or
`module` instance.  This uses the existing cache data structure
attached to a function in a new way so these caches must be
cleared on a miss in order to prepare the cache for use with
`LOAD_ATTR_CACHED`.

This also rewrites two uses of `_type` to an equivalent
`_object_type_getattr` where the receiver type is know to be
megamorphic.  This saves some churning of the caches.

Based on Facebook D18368800