Do not cache STORE_ATTR when receiver layout changes
Previously, we would add a PIC entry using the layout ID of the
receiver after a `STORE_ATTR`.  If the `STORE_ATTR` updated the layout
ID than than a bogus entry would appear in the PIC.

This works around the problem by only updating the cache when the
layout ID of the receiver before the `STORE_ATTR` matches the layout
ID of the receiver after the `STORE_ATTR`.

Based on Facebook D18494238