Monomorphic caching for STORE_ATTR for writing attributes in-object and updating layout
This change introduces specialized opcodes STORE_ATTR_INSTANCE_UPDATE for monomorphic forms of `STORE_ATTR` where attribute values are directly written into object instances & layout is updated.  This uses the existing cache data structure attached to a function in a new way so these caches must be cleared on a miss in order to prepare the cache for use with `STORE_ATTR_CACHED`.

**Perf Impact (CG for 5th request on django_minimal)**

- Base: 4485677
- This diff: 4225055
- Improvement: -5.81%

Based on Facebook D18495977