Rewrite printf format strings with %d to f-strings
Summary: This adds support for rewriting printf-format strings containing `%d` (and the aliases `%i`, `%u`). The trick here is that we need to perform some conversion to an int before formatting the final value, so we add a helper method to the `str` class that performs this conversion.

Based on Facebook D18408811