Support rewriting printf format strings without tuple on RHS
This adds support for rewriting printf-style formatting into f-string when the right hand side is not a tuple. This is only allowed for strings with a single specifier at the moment, because we can then inject a conversion function that checks whether the supplied value is a 1-value tuple or if it is not creates one.

i.e. `"%s" % x` is rewritten to `f"{_mod_check_single_arg(x)[0]}"`

Based on Facebook D18408810