Rewrite printf string format into f-strings in compiler
Add an optimization to the compiler that rewrites some printf-style formatting (i.e. `"%s;%s" % (a, b)`) into equivalent f-strings (i.e. `f"{a};{b}"`).

This is rewrite is performed on the AST before emitting bytecode.

This is a first version that just supports format strings containins `%%`, `%s`, `%r`, `%a`.

Based on Facebook D18408813