Make generic dumper robust against layout-ids that have no associated type
Summary: We have in invariant in pyro that the we have a valid layout present in the runtime layout table for each object, and each of these layouts is connected to a type. It is however convenient to not rely on it in the debug dumper so we do not just crash in the presence of bugs that violate the invariant.

Based on Facebook D18581950