Implement PySys_SetArgv and PySys_SetArgvEx
Implement these two functions as part of a larger overhaul to rewrite `main()` in terms of the CPython C-API. Most of the tricky path manipulation has been lifted into Python.

Note on tests: CPython will pass a pointer to the middle of the original argv array, so the test code duplicates this behavior.

Based on Facebook D18410142