Fix refcounting in PyArena
(port of D18518095)

`PyArena_AddPyObject` is an internal compiler API that guarantees that a `PyObject*`
will be kept alive as long as the arena is kept alive, so the compiler can
(ab)use borrowed references with abandon. There's one quirk with this
in our runtime: we can't guarantee that we keep the handle alive if all
the references to it are borrowed. [PyPy suffers](https://cython.readthedocs.io/en/latest/src/userguide/pypy.html) from this problem as well.

The solution to this problem, presented in this diff, is to let the caller keep
its owned reference and instead DECREF all the `PyObject*`s in the arena
at arena destruction time.

Based on Facebook D18483941