Have a fast-path in str.encode for ASCII-only inputs
The majority (or entirety so far) use-case of `str.encode` is to encode a string in `strict` mode, using either the `utf-8`, `latin-1`, or `ascii` encoding. Since almost all encodings map `ascii` bytes to `ascii`, we can have a fast-path that simply checks if a string is `ascii`, and just copies the string over into a buffer and immediately returns the buffer.

This fast-path covers every call to `str.encode` in the `django 100` benchmark.

Based on Facebook D18620709