Fix GC issue in ApiHandle::ensure
If a GC happened in between `ApiHandle::alloc` and
`dictAtPutIdentityEquals`, the `ApiHandle` would contain a dangling reference
to a moved pointer. This fixes that by making `::ensure()` own a reference
to the handle while inserting into the table.

It also inlines `ApiHandle::alloc` into its only caller, `ApiHandle::ensure`.

Based on Facebook D18407167