Implement memoryview.__setitem__ for integer indexing
Implements `memoryview.__setitem__` similarly to CPython's `memory_ass_sub`.

The function first assumes that `key` is a base `Int`, and that `value` is the appropriate type for the `memoryview`'s `format`. If either assumption is broken, it returns `_Unbound`, and some conversions are done in managed code.

Most of the functionality is found in `packObject`, which writes to the underlying memory of the `memoryobject`. This means that we pass around `address()`, and that `packObject` should therefore avoid doing any allocations.

`packObject` simply checks the `format` of the `memoryview` object, makes sure that `value` is of an appropriate type for that `format`, and then writes the object at the index specified in `byte_index`.

Based on Facebook D17533539