Re-implement _Py_DecodeUTF8_surrogateescape without Runtime
Summary: When I implemented this, I didn't realize that it has to be available before the runtime starts up (because it is used to parse command-line args). @[662308721:Carl Shapiro] and I decided to lock this to support ASCII for now, as it probably doesn't need to support UTF-8 args internally (especially as this function is applicable to Mac OS X only) and there is more important stuff to work on.

Based on Facebook D18639779