Simplify string hashing in type attribute dictionary
Summary: Type attributes are only accessed by interned strings. This allows to slightly simplify the hash retrieval because for interned strings we know a-priori that the hash value is already computed. This also allows us to get by without having a `thread` pointer around in some cases simplifying calls.

Based on Facebook D18667124