Make C-API method trampolines own references to arguments
The C-API method trampolines were inappropriately using borrowed references to store `PyObject*`
that needed to last for the duration of the trampoline. This is wrong because
the trampolines run arbitrary code, including potentially a GC. The trampolines
instead need to own the references. This diff changes the references to owned
references and decrefs after the call is over.

This fixes some ASAN issues that were popping up trying to run Django on
master.

Based on Facebook D18599441