Enable eval and exec with globals dict
Pyro's implementation of `exec` and `eval` is implemented to only work with module dictionaries. However, CPython allows the use of arbitrary dictionaries as a globals parameter. The Cython compiler uses this extensively to generate the *.c files.

To allow this, we create a temporary module object that gets updates with the globals dict. Once the execution of the code happens, it updates back the globals dictionary.

Based on Facebook D18761244