Place default attribute getter/setter into typeslot where possible
It seems to be common in cpython to fill the tp_getattro typeslot with PyObject_GenericGetattr, or the tp_setattr typeslot with PyObject_GenericSetAttr. These have the exact same effects as `object.__getattribute__` and `object.__setattr__` but would have different identities and our caching strategies would not trigger.

Change the extension slot init code to replace them with direct reference to `object.__getattribute__` and `object.__setattr__`.

Based on Facebook D18739093