Remove putenv usage
`putenv` does not play nice. It stores the pointers it is given straight
into the environment. This is exceedingly tricky to manage with our moving
GC, so instead we don't bother. Fortunately, CPython offers knobs to turn off
`putenv` if the host system does not support it.

Based on Facebook D18515937