Implement set.clear
Summary: Create a C function to clear a set by accessing the internal bucket and filling it up with NoneTypes and setting the set size to zero. Note that this was specifically added to `SetBuiltins` and not `SetBaseBuiltins` to avoid having frozen set inherit this implementation.

Based on Facebook D18769194