Don't store Modules directly in cache tuples
Two of the shutdown-time memory leaks I'm currently looking at are
partially caused by cache tuples keeping modules alive past when they'd
naturally die. To avoid this, give every module a small integral ID at creation
time (similar to `LayoutId`), and use it as the cache key. This adds a few
instructions to the cache hit path, which is ~0.04% in cg_instructions on
django100 (and in the noise for time).

The ID is stored in the hash code field of each `Module`. This means
`hash(module)` will return its ID. In the unlikely event that this is a
problem, there are a number of things we can do to provide a stronger hash
code.

Also: Change the `DCHECK()` in `reserveLayoutId()` to a `CHECK()`, to match the
`CHECK()` I added in `reserveModuleId(). Overflowing either of these counters
is a big deal, and neither function should be called very often outside of
warmup.

Based on Facebook D18647986