Use CALL_FUNCTION instead of CALL_FUNCTION_EX in type.__call__
The current implementation of `type.__call__` always uses
`Interpreter::callEx` even when there are no keyword arguments
passed in the call.  This is inefficient since it requires
arguments to be boxed.

This change lets it selectively use `Interpreter::call` when
`kwargs` is empty for `type.__call__` calls, since a `call`
is much faster than a `callEx`.

Based on Facebook D18819165