Add _bytes_decode and add more calls to _str_encode
Adds a fast-path for decoding `bytes` that are all `ASCII`. If the object is not exactly a `Bytes` object, or if it's not all `ASCII`, or if the function is encoding anything other than `ASCII`, `UTF-8`, or `Latin-1`, the fast-path will return `Unbound`.

With this diff in, we only have 4 more requests that go through a `_codecs` search-path per-request, which are all from `_io` runnning `_codecs.getincremental{de|en}coder`, which then feeds into `_codecs.lookup`, and results in a cached value immediately being returned.
There are also a few calls to `ascii_encode` being made directly through writing to some `_io` objects that have been sped up by this diff.

Based on Facebook D18690384