Avoid callbacks and reflection in common structure sequence operations
Previously, on every get and set the bounds were read from
attributes in the type of the structure sequence, which require
loading the instance type and reading out of its dictionary.
However, these same values can be derived from the structure
sequence instance or its underlying tuple instance.  Also, every
get and new required a callbacks into a Python function.
However, the effect of the Python function could be achieved with
ordinary runtime calls.

This also switches the attribute of a structure sequence that are not
in the underlying tuple to be stored in in-object attributes instead
of overflow attributes.  This simplifies the bounds computation and
reduces the time- and space-overhead of the initializing assignments
to the newly created structure sequence.

Based on Facebook D18815890