Implement memoryview.__setitem__ with slices
Since setting slices is already on the slow-path, `_memoryview_setslice` doesn't make any assumptions like `_memoryview_setitem` does. It does have fast-paths when `step == 1` however.

If `memoryview.__setitem__` is called with a `Slice`, and the passed in object is either `Bytes` or a `ByteArray`, than the `memoryview` must be of format `B`, which means we can index and write to the address of the `memoryview` as if it were a `byte*`.

If it is called with another `memoryview` object, we utilize `unpackObject` to do most of the work of reading out values from it when taking into account its `format`.

Based on Facebook D17815636