Properly support Module subclasses in `doLoadAttrModule()`
The old code would try to use `Module::cast()` with an object that
could be a `Module` subclass, which isn't safe. Since the previous line just
verified the type, we can use `rawCast()` to skip the check inside `cast()` and
avoid the overhead of a `HandleScope` and `Handle<RawModule>`.

Based on Facebook D18845738