Move main interpreter code to Py_BytesMain()
Move interpreter initialization and script running to `Py_BytesMain()`, to conform to C-API. Note that `Py_BytesMain()` was added in 3.7 but we use it here (and wrap `Py_Main()` around it) to make things a little cleaner.

As part of this, we use `PySys_SetArgv()` instead of our custom function (now that we have `wchar_t** argv` available). Most of the code in Modules/main.cpp is just moved, with a few small modifications, which I call out below.

Based on Facebook D18639776