Match CPython command-line argument handling
Since we're trying to be a drop-in replacement for CPython, we should handle all command-line arguments the way that CPython does, or raise an exception if we don't handle them. Some of these options might work correctly as of today, but it will be left to future diffs to verify that and enable them.

This code matches the structure and ordering of CPython's init routine closely, so we can easily verify where we are different.

One change: We opt to not handle arguments which have been deprecated and which CPython silently ignores; we will abort and print a help message.

Based on Facebook D18639777