Save progress made in repeated calls to Scavenger::processGrayObjects()
This allows repeated calls to `processGrayObjects()` without
reprocessing the whole heap, which avoids both a) wasting CPU cycles processing
objects that are known to be black, and b) adding `WeakRef`s to the processing
queue multiple times, which corrupts the list.

We currently only have one call to `processGrayObjects()` per collection, but I
need to add another one after `processFinalizableReferences()` to fix the issue
in T58693275.

Based on Facebook D18821213