Verify heap after every GC in debug builds, fix existing failures
We already had `Heap::verify()` but weren't calling it anywhere. Call
it as the final step in a garbage collection to help catch bugs early.

This caused a bunch of test failures because `processNativeList()` can create
gray objects (the specific one I debugged was a `NotIn` object from the
compiler, which has an overflow tuple). When based on D18821213, we can fix all
of the failures with a single call to `processGrayObjects()` after
`processFinalizableReferences()`.

I also added a check to `enqueueReference()` that we're not trying to enqueue
an object already in a queue, which is how this diff fails without D18821213.

Based on Facebook D18839381