Remove CPython source files from generate_cpython_sources.py
`generate_cpython_sources.py` does not modify any implementation (*.c) file anymore. This only modifies the cpython headers using the pyro headers as source.

This updates `CMakeLists.txt` and `CPythonBuild.txt` to split up the cpython headers and sources to draw a cleaner distinction of what needs to be modified. This is in preparation to allow exporting the headers to a build directory which can be used to build third-party C Extensions.

Based on Facebook D18892743