Remove un-needed StringIO methods
Since the `StringIO` implementation was copied from `TextIOWrapper`, there are a lot of methods and fields that are un-needed. This diff removes the ones that should be no-ops.

One of the properties this removes is `buffer`, so the diff also makes sure that all the previous callers of `buffer` call `_buffer` instead.

Based on Facebook D18973441