Fix a bug in removeDeadWeakLinks that wrongly typecasts None to WeakLink
Previously, the code assumed that the current node in the linked list traversal was never the head node.  However, this never worked for the case when the head node is found to need removal.  This caused a cast to fail which was intermittently observed in Django.

This change fixes the head removal case and adds extra test coverage.

Based on Facebook D18983313