Delete CPythonBuild.cmake
Summary: There's no need for this to be split up into a different cmake file anymore. This keeps the entire build in a single CMakeLists.txt file. Furthermore, this renamed the `cpython` library target to `extension-modules` as this is really what we are compiling from that library.

Based on Facebook D18964371